requirements.txt
flask==1.0.3
google-cloud-language==1.2.0
google-cloud-firestore==1.4.0
tweepy==3.7.0
gunicorn==19.9.0
